Adjusted and documented class names and for disabling http -> https redirection
While implementing the ability to prevent listening on port 80 I noticed this
documentation was missing so I added it.

Based on feedback from the PR providing capability to disable listening on port
80 these related settings to avoid redirecting from http -> https were updated
to be more clear.

Warnings were also introduced to highlight the case of using the older less
understandable value and redirection to http is automatically disabled if plain
http is disabled.

Ticket: ENT-2071
Changelog: None