Changed default self upgrade target version to be that of Hubs binary version
As a general rule MPF => Hub Binary Version >= Client Binary Versions.
Previously the default self upgrade binary version targeted the version of the
MPF release. This causes hosts to target a new binary release possibly before
the hub binary is upgraded, and is annoying for anyone tracking LTS branches
that are not necessarily on a specific release.

With this change the hub shares its binary version with clients and clients use
that binary version as their default target.

Ticket: ENT-10664
Changelog: Title