Added bundle to facilitate migration of ignore_interfaces.rx from inputdir to workdir
This change aims to increase awareness of the change in behavior for the
location of the ignore_interfaces.rx file. It emits reports about the presence
and status of the file.

In order to further facilitate the migration, this policy will copy the file
from inputdir to workdir if the class
default:mpf_auto_migrate_ignore_interfaces_rx_to_workdir is defined and the
file is present.

Ticket: ENT-9402
Changelog: Title
(cherry picked from commit 249f3805bcb5b3fd0acec524d3b07c226a6247ba)