Fixed restoration of Mission Portal application to packaged content when modified
Prior to this change if Mission Portal application code was modified it would
not be returned to the packaged state because the deployed code would have a
newer timestamp than the packaged source. This change switches to using binary
comparison of Mission Portal to it's packaged source to determine if the
deployed application should be reverted.

Ticket: ENT-10962
Changelog: Title