Fixed policy compatibility with 3.15.x, disabled aws ec2 inventory for versions less than 3.18.0
This change aims to fix policy compatibility between 3.18.x and 3.15.x by
guarding the policy with a macro to prevent agents older than 3.18.0 from
evaluating policy which uses features the agent is not capable of.

While 3.15.x has been out of support since December 31st 2022 we strive to minimize
any pain in upgrading. The change that introduced this issue extended ec2
metadata to support and prefer imdsv2 uses a function, validjson() and an
attribute, content which were first introduced in 3.16.0.

Ticket: ENT-11178
Changelog: Title