Added ability to influence default package manager and inventory via Augments
This change adds the ability to override the default package manager defined by
the MPF for a platform via Augments based on keys expected to match the value of
sys.flavor.

For example, this would set the default package manager to yum on AIX  7 and
aptitude on Ubuntu 20.

{
  "variables": {
    "default:def.override_default_package_manager": {
      "value": {
        "ubuntu_20": "aptitude",
        "aix_7": "yum"
      },
        "comment": "This variable provides the ability to override the default package manager to use for a platform. Keys are based on the value of $(sys.flavor) for the targeted platform."
    }
  }
}

This change also adds the ability to specify package managers that should be
inventoried by default (in addition to the default package manager).

For example, augments containing a definition like this would inventory snap and
flatpack in addition to the default apt_get on Ubuntu 20, and yum in addition to
the default nimclient on AIX 7.

{
  "variables": {
    "default:def.additional_package_inventory": {
        "value": {
            "ubuntu_20": [ "snap", "flatpak" ],
            "aix_7": [ "yum" ]
        },
        "comment": "This variable provides the ability to extend the default package managers to inventory for a platform. Keys are based on the value of $(sys.flavor) for the targeted platform."
      }
  }
}

Ticket: CFE-3612
Changelog: Title