Freebsd service management now uses one prefixed service commands
On FreeBSD service commands only work if a service has been enabled in
/etc/rc.conf. If it has not been enabled then a services promise with a
service_policy of "start" or "stop" results in a failure to change the system
state during each agent execution. Switching to "one" prefixed commands, e.g.
onestatus, onestop, and onestart allows services that are not enabled to still
be managed and does not interfere with services that have been enabled in
/etc/rc.conf.

Ticket: CFE-4323
Changelog: Title