Made AIX watchdog ignore stale pidfiles
Ticket: CFE-4335
Changelog: AIX watchdog now handles stale pids

Code uses /proc opportunistically only if it exists; this improves
accuracy for systems with /proc and doesn't affect other systems.
(Turns out this doesn't matter as the template is only for AIX which
does have /proc)

This leaves in place a race condition if many watchdog processes are
started in very short succession, since the pidfile is not atomically
checked and updated.  For purposes of the watchdog this is probably good
enough.  (See https://stackoverflow.com/a/688365/5419599 for more on
this.)

(cherry picked from commit f53436b11d12573ec7c49dc677c68fbecbcfc2d3)