Added missing trailing slash to .no-distrib directory in MPF
I believe this has no functional difference, but it's an important hint at least
to the policy reader that it is a directory.

Unlike other access rules I left off the condition to only add the access rule
if the directory exists so that if it is created there will not be a window of
time for which clients would be able to get files from the directory.

If the directory does not exist there are info level logs that are emitted, but
that seems acceptable since we do not log info from cf-serverd by default. There
is apparently no issue making an access promise for a directory that does not
exist, I see it listed in the cf-serverd access summary from verbose logging.

Ticket: CFE-4453
Changelog: None