Modified several common bundles to be agent bundles to limit access in SELinux policy
This should simplify selinux policy where we want the agent to have broad access but other components like serverd, execd, hub to have more limited access.

The paths bundle checks for the existence of many commands and this operation can cause SELinux AVCs.
The inventory bundle uses the curl command to see if the host is an AWS instance and if so collects some inventory from a well known API/IP.
The cfe_internal_hub_vars needs to determine if php-fpm executable is present.

As far as I can tell there isn't a need for the vars/classes defined in paths and inventory in components other than agent so this change should be OK.

Ticket: ENT-12954
Changelog: title

more common to agent fixes