Prevented nfs server inventory from doing unnecessary extra work
We have received reports of slow policy execution on hosts with many
nfs mounts (hundreds). Part of this expense has to do with the fact that the
promise re-defined the itself on each pass of the policy.

This change prevents that additional un-necessary processing and re-definition
by restricting the promise only to when inventory_linux.nfs_servers is not
already defined. This single change reduced processing time in one case by ~70
seconds (from ~90 seconds to ~20 seconds).

Ticket: ENT-13210
Changelog: Title