Stopped enforcing permissions of public docroot scripts
This change stops enforcing permissions of files in the scripts subdirectory of
the public docroot. The volume of files in this directory do not make it
sensible to check permissions during each agent execution.

To improve performance of policy runs on hubs this specific enforcement is
skipped unless a class is set. These files permission should be fully handled by
the package.

Ticket: ENT-13243
Changelog: Title