Fixed cfruncommand for Windows causing "Too many arguments" error
The issue was that the entire command line was sent and the '&' which should be a command separator was instead being interpreted as another argument.

This caused the argument parser to try and interpret '&' as the filename as in this special syntax: "cf-agent <policy-file>" and since there is more text after '&' in the existing cfruncommand value the "Too many arguments" error is caused.

Ticket: ENT-13530
Changelog: title
(cherry picked from commit d9238b24b833cc8163b3325f24b26bfa11593b40)